Bugfix: wrong database dropped in the MongoTest.tearDown() method
In the MongoTest.tearDown() method, the 'test' database was dropped
instead of the 'test_db' one.

This caused the MongoTest.test_insertion_mongo() test to fail when
the test suite was ran more that once.
As no database was dropped, some ancient tweets were still stored,
which broke the assertEqual(size_db, nb_fetched_tweets).

Signed-off-by: Balthazar Rouberol <rouberol.b@gmail.com>